♻️ matchingしている投稿の取得する処理の共通化
