+++ src/graphql/logics/index.ts
@@ -1,0 +1,1 @@
+export { matchPosts } from './matchPosts'
